+++ src/main/java/com/wesovilabs/workshops/graphql/database/repository/ActorRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ActorRepository extends JpaRepository<ActorEntity,Long> {
+public interface ActorRepository extends JpaRepository<ActorEntity,Integer> {
 }
